+++ src/util/mount.c
@@ -13,6 +13,7 @@
  * 
 */
 
+#define _GNU_SOURCE
 #include <errno.h>
 #include <fcntl.h>
 #include <stdio.h>
@@ -23,6 +24,7 @@
 #include <unistd.h>
 #include <stdlib.h>
 #include <limits.h>
+#include <libgen.h>
 
 #include "config.h"
 #include "util/file.h"
@@ -36,10 +38,42 @@ int singularity_mount(const char *source, const char *target,
                       const char *filesystemtype, unsigned long mountflags,
                       const void *data) {
     int ret;
+    int mount_errno;
     uid_t fsuid = 0;
+    char dest[PATH_MAX];
+    char *realdest;
+    int target_fd = open(target, O_RDONLY);
+
+    if ( target_fd < 0 ) {
+        singularity_message(ERROR, "Target %s doesn't exist\n", target);
+        ABORT(255);
+    }
+
+    if ( snprintf(dest, PATH_MAX-1, "/proc/self/fd/%d", target_fd) < 0 ) {
+        singularity_message(ERROR, "Failed to determine path for target file descriptor\n");
+        ABORT(255);
+    }
 
     if ( ( mountflags & MS_BIND ) ) {
         fsuid = singularity_priv_getuid();
+    }
+
+    realdest = realpath(dest, NULL); // Flawfinder: ignore
+    if ( realdest == NULL ) {
+        singularity_message(ERROR, "Failed to get real path of %s %s\n", target, dest);
+        ABORT(255);
+    }
+
+    if ( (mountflags & MS_PRIVATE) == 0 && (mountflags & MS_SLAVE) == 0 ) {
+        if ( strncmp(realdest, CONTAINER_MOUNTDIR, strlen(CONTAINER_MOUNTDIR)) != 0 &&
+             strncmp(realdest, CONTAINER_FINALDIR, strlen(CONTAINER_FINALDIR)) != 0 &&
+             strncmp(realdest, CONTAINER_OVERLAY, strlen(CONTAINER_OVERLAY)) != 0 &&
+             strncmp(realdest, SESSIONDIR, strlen(SESSIONDIR)) != 0 ) {
+            singularity_message(VERBOSE, "Ignored, try to mount %s outside of container %s\n", target, realdest);
+            free(realdest);
+            close(target_fd);
+            return(0);
+        }
     }
 
     /* don't modify user groups */
@@ -51,12 +85,19 @@ int singularity_mount(const char *source, const char *target,
         /* NFS root_squash option set uid 0 to nobody, force use of real user ID */
         setfsuid(fsuid);
     }
-    ret = mount(source, target, filesystemtype, mountflags, data);
+
+    ret = mount(source, dest, filesystemtype, mountflags, data);
+    mount_errno = errno;
+
+    close(target_fd);
+    free(realdest);
+
     if ( singularity_priv_userns_enabled() == 0 && seteuid(singularity_priv_getuid()) < 0 ) {
         singularity_message(ERROR, "Failed to drop privileges: %s\n", strerror(errno));
         ABORT(255);
     }
 
+    errno = mount_errno;
     return ret;
 }
 
@@ -67,6 +108,7 @@ int check_mounted(char *mountpoint) {
     char *rootfs_dir = CONTAINER_FINALDIR;
     unsigned int mountpoint_len = strlength(mountpoint, PATH_MAX);
     char *real_mountpoint;
+    char procmounts[PATH_MAX];
 
     singularity_message(DEBUG, "Opening /proc/mounts\n");
     if ( ( mounts = fopen("/proc/mounts", "r") ) == NULL ) { // Flawfinder: ignore
@@ -79,10 +121,20 @@ int check_mounted(char *mountpoint) {
         mountpoint[mountpoint_len-1] = '\0';
     }
 
-    real_mountpoint = realpath(mountpoint, NULL); // Flawfinder: ignore
+    real_mountpoint = realpath(joinpath(rootfs_dir, mountpoint), NULL); // Flawfinder: ignore
     if ( real_mountpoint == NULL ) {
-        // mountpoint doesn't exist
+        // mountpoint doesn't exists
         return(retval);
+    }
+
+    if ( snprintf(procmounts, PATH_MAX-1, "%s/proc/%d/mounts", rootfs_dir, getpid()) < 0 ) {
+        singularity_message(ERROR, "Can't construct path %s/proc/%d/mounts\n", rootfs_dir, getpid());
+        ABORT(255);
+    }
+
+    if ( strcmp(real_mountpoint, procmounts) == 0 ) {
+        singularity_message(ERROR, "Attempt to override /proc/mounts, aborting\n");
+        ABORT(255);
     }
 
     singularity_message(DEBUG, "Iterating through /proc/mounts\n");
@@ -91,7 +143,7 @@ int check_mounted(char *mountpoint) {
         char *mount = strtok(NULL, " ");
 
         // Check to see if mountpoint is already mounted
-        if ( strcmp(joinpath(rootfs_dir, real_mountpoint), mount) == 0 ) {
+        if ( strcmp(real_mountpoint, mount) == 0 ) {
             singularity_message(DEBUG, "Mountpoint is already mounted: %s\n", mountpoint);
             retval = 1;
             break;